Use `Instances` for sending NOTIFYs
